fix: move screens and widgets out of util folders
