+++ lib/card_templates/barcode_scanner_util.dart
