+++ lib/card_templates/contact_card_form.dart
@@ -12,7 +12,7 @@ import 'package:magicepaperapp/l10n/app_localizations.dart';
 import 'package:magicepaperapp/provider/getitlocator.dart';
 import 'package:magicepaperapp/native_canvas/native_canvas_editor.dart';
 import 'package:magicepaperapp/utils/page_route_util.dart';
-import 'package:magicepaperapp/card_templates/utils/barcode_scanner_util.dart';
+import 'package:magicepaperapp/card_templates/barcode_scanner_util.dart';
 import 'package:magicepaperapp/view/widgets/common_scaffold_widget.dart';
 
 AppLocalizations get appLocalizations => getIt.get<AppLocalizations>();
