+++ lib/card_templates/utils/image_picker_util.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:magicepaperapp/utils/image_crop_screen.dart';
+import 'package:magicepaperapp/view/image_crop_screen.dart';
 import 'package:magicepaperapp/utils/image_source_picker.dart';
 import 'package:path_provider/path_provider.dart';
